client/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
albucore==0.0.20
albumentations==1.4.21
annotated-types==0.7.0
attrs==24.2.0
blinker==1.9.0
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cuda-python==12.6.0
cycler==0.12.1
eval_type_backport==0.2.0
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.9.0
gevent==24.10.3
geventhttpclient==2.0.2
greenlet==3.1.1
grpcio==1.67.1
humanfriendly==10.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.1
matplotlib==3.9.2
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.2
multidict==6.1.0
networkx==3.4.1
numpy==1.26.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnx==1.17.0
onnxruntime==1.19.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
protobuf==4.25.5
psutil==6.1.0
py-cpuinfo==9.0.0
PyAutoGUI==0.9.54
pycairo==1.27.0
pydantic==2.10.2
pydantic_core==2.27.1
PyGetWindow==0.0.9
PyGObject==3.50.0
PyMsgBox==1.0.9
pyparsing==3.2.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-rapidjson==1.20
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.45
scipy==1.14.1
screeninfo==0.8.1
seaborn==0.13.2
setuptools==75.1.0
simsimd==6.2.1
six==1.16.0
stringzilla==3.11.0
sympy==1.13.1
thop==0.1.1.post2209072238
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.6
triton==3.1.0
tritonclient==2.51.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.0.196
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.17.0
zope.event==5.0
zope.interface==7.1.1
